Inventory system and stamina
So the inventory system is on the player, its a script called inventory, which manages the ui object (attached to player) and the actual data of the players invetory. Items in the game are scriptable objects, so they can be made inside the editor, they belong to containers, which are an object which can be instantiated into the world, call the players inventory system, pass the item then kill themselves.

The one kind of tricky buisness to all of this is the UI unity uses which is not great. i had to manually give all the buttons on the ui manager individual string arguments to pass to the inventory script - but that is what the select inventory slot is - its being called by the onclick, which passes basically a button index to the inventory class.

all the rest is easy peasy unity stuff