inventory system now deals with equipment, schedule now sets enemy pos to where they should be
updated the inventory sytem so that it works with equipment, the equipment sets values on a collection of mesh filters attached to the player object, and then the interact button takes the currently equipped sword into acount. we have slots for differnt parts of the body, the inventory screeen updates to show you equipped weapons.

Stamina now runs down, when it is empty you will wake up the 'next day' at a preset time, atm it doesnt advance the day, but it does set the ai to the place they should be